backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
nltk==3.9.3
numpy==2.4.2
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
regex==2026.2.28
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
vaderSentiment==3.3.2
sqlalchemy
psycopg2-binary
python-dotenv
groq
passlib[bcrypt]
python-jose
